+++ Src/FastData/Contexts/HashSetPerfectContext.cs
@@ -2,8 +2,8 @@ using Genbox.FastData.Abstracts;
 
 namespace Genbox.FastData.Contexts;
 
-public class HashSetPerfectContext<T>(KeyValuePair<T, uint>[] data, uint seed) : IContext
+public class HashSetPerfectContext<T>(KeyValuePair<T, ulong>[] data, uint seed) : IContext
 {
-    public KeyValuePair<T, uint>[] Data { get; } = data;
+    public KeyValuePair<T, ulong>[] Data { get; } = data;
     public uint Seed { get; } = seed;
 }
