+++ Src/FastData/Contexts/Misc/HashSetEntry.cs
@@ -3,4 +3,4 @@ using System.Runtime.InteropServices;
 namespace Genbox.FastData.Contexts.Misc;
 
 [StructLayout(LayoutKind.Auto)]
-public record struct HashSetEntry<T>(uint Hash, int Next, T Value);
+public record struct HashSetEntry<T>(ulong Hash, int Next, T Value);
